refactor: move server config to top level
