+++ rigflow-server/src/server/mod.rs
@@ -1,5 +1,4 @@
 pub mod app_state;
-pub mod config;
 pub mod discovery;
 pub mod radio_manager;
 pub mod radio_api;
