+++ rigflow-server/src/server/radio_manager.rs
@@ -7,7 +7,7 @@ use tokio::task::JoinHandle;
 
 use rigflow_core::radio::{LeaseId, RadioDescriptor, RadioId};
 
-use crate::server::config::ServerConfig;
+use crate::config::ServerConfig;
 use crate::server::radio_types::{
     AcquireRadioResult, AcquireRequest, ClientId, LeaseRecord, RadioManagerConfig,
     RadioManagerError, RadioState, RadioSummary, StopReason, WorkerCommand, WorkerExit,
